Replace libc in Linux with rustix
rustix is a crate which provides I/O-safe wrappers around raw system
calls on Linux and libc on other Unixes. This PR makes it so the
functions in this crate call the rustix functions.

This has a number of benefits. It removes a substantial amount of
unsafe code from this crate. In addition, raw syscalls are often more
efficient than libc calls; sometimes they can even be inlined into
their calling functions.

Two compromises were made in this PR:

- This crate does not have I/O safety yet, while rustix does. So I
  borrow the raw file descriptor for now to avoid a breaking change.
- The BSD equivalents to the Xattr functions are not yet exposed in
  Rustix. So I still use libc in the BSD backend.

Closes #40

Signed-off-by: John Nunley <dev@notgull.net>