fix: freebsd's xattr returns the number of bytes read
On linux, the xattr API returns the ERANGE if the result doesn't fit. On
FreeBSD, it returns the number of bytes read.

Previously, this wasn't an obvious issue because we always checked the
size before we read the attribute. It was still incorrect, but would
have required a race to trigger a bug.

Now, it's a more pressing issue because we try reading once with a
4KiB buffer before we check the size.

The fix is simple: make the target buffer 1 byte larger than it has to
be and make sure we always read less than the size of the buffer we
passed in.