Exclude a script file from the published package (#88)
During a dependency review we noticed that the xattr crate includes a ci script file. This file is not required for building xattr as depndency. The script file might at some point become problematic. As of now they prevent any downstream user from enabling the `[bans.build.interpreted]` option of cargo deny.

I opted for using an explicit include list instead of an exclude list to prevent these files from beeing included in the published packages to make sure that everything that's included is an conscious choice.